requirements.txt
openenv-core[cli]>=0.2.0
pydantic>=2.0
numpy>=1.24
python-dotenv>=1.0
openai>=1.0
gradio>=4.0